refactor(statistics): extract comparison chart Y-axis tick to its own file
Moved ComparisonYAxisTick and its layout constants out of
ComparisonChart.tsx into ComparisonYAxisTick.tsx. No behavior change.
Per review feedback.
